+++ project/assets_cartographer/templatetags/assets.py
@@ -3,9 +3,10 @@ Assets tags
 
 TODO: another tag for loading asset kind that are not stylesheet or javascript
 """
+from django.template.defaulttags import register
+
 from project.assets_cartographer import manifest
 from project.assets_cartographer.parser import AssetTagsManagerFromManifest
-from django.template.defaulttags import register
 
 def render_asset_tags(kind, *bundle_names):
     """
@@ -19,6 +20,25 @@ def render_asset_tags(kind, *bundle_names):
 
 
 @register.simple_tag
+def asset_tag(kind, *bundle_names):
+    """
+    Build tag for any kind of asset
+    
+    Usage
+    *****
+    
+    Give asset kind name as first argument, then package names (one or more) as other arguments: ::
+    
+        {% asset_tag "stylesheet" "css/item1.min.css" "css/item2.min.css" .. %}
+        
+    Depending on settings.ASSETS_PACKAGED it would build an unique tag (if True) for the packaged asset file or tag (if False) for each package components files.
+    
+    Obviously, the given kind have to exist in your asset manifest.
+    """
+    return render_asset_tags(kind, *bundle_names)
+
+
+@register.simple_tag
 def stylesheet_tag(*bundle_names):
     """
     Build tag for stylesheet assets
@@ -26,7 +46,7 @@ def stylesheet_tag(*bundle_names):
     Usage
     *****
     
-    Just gives package names (one or more) as arguments: ::
+    Just give package names (one or more) as arguments: ::
     
         {% stylesheet_tag "css/item1.min.css" "css/item2.min.css" .. %}
         
@@ -43,7 +63,7 @@ def javascript_tag(*bundle_names):
     Usage
     *****
     
-    Just gives package names (one or more) as arguments: ::
+    Just give package names (one or more) as arguments: ::
     
         {% javascript_tag "css/item1.min.js" "css/item2.min.js" .. %}