+++ project/manager_frontend/forms/systems.py
@@ -9,10 +9,14 @@ from django import forms
 from django.utils.translation import ugettext_lazy as _
 from django.core.files.storage import FileSystemStorage
 
+from project.recalbox_manifest import manifest as RECALBOX_MANIFEST
+
 from project.manager_frontend.forms import CrispyFormMixin
+
 from project.utils.imports import safe_import_module
 
-class SystemCreateForm(CrispyFormMixin, forms.Form):
+#class SystemCreateForm(CrispyFormMixin, forms.Form):
+class SystemCreateForm(forms.Form):
     """
     Create a new system directory
     """
@@ -23,7 +27,7 @@ class SystemCreateForm(CrispyFormMixin, forms.Form):
         self.available_systems = kwargs.pop('available_systems')
         
         super(SystemCreateForm, self).__init__(*args, **kwargs)
-        super(forms.Form, self).__init__(*args, **kwargs)
+        #super(forms.Form, self).__init__(*args, **kwargs)
         
         self.fields['name'] = forms.ChoiceField(label=_('Add a new system'), choices=self.available_systems, required=True)
     
@@ -32,4 +36,4 @@ class SystemCreateForm(CrispyFormMixin, forms.Form):
         
         os.mkdir(os.path.join(settings.RECALBOX_ROMS_PATH, name), 0755)
         
-        return settings.RECALBOX_MANIFEST[name]
+        return RECALBOX_MANIFEST[name]