+++ project/manager_frontend/views/logs.py
@@ -2,7 +2,6 @@
 Views for logs files
 """
 from django.conf import settings
-from django.shortcuts import render
 from django.views.generic import TemplateView
 
 class LogsView(TemplateView):