+++ project/recalbox_manifest/registry.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""
+Recalbox manifest registry
+"""
+# Just a dummy dict waiting to be filled
+manifest = {}